Add files via upload
This project is a web application that allows users to upload an image and convert it into either a watercolor sketch or a pencil sketch. The app leverages OpenCV for image processing and Streamlit for the user interface. Users can choose their preferred sketch style and download the resulting sketch.